Extract issue transformations from transformation interface
